task 2

+++ 0x0F-function_pointers/2-int_index.c
@@ -12,7 +12,7 @@
  */
 int int_index(int *array, int size, int (*cmp)(int))
 {
-	unsigned long int i, j = 0;
+	int i, j = 0;
 	int k;
 
 	if (size <= 0)
